requirements.txt
fastapi==0.111.2
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
pydantic==2.5.1
python-jose==3.3.0

python-dotenv
